01_Miranda {modelfree}	R Documentation

Example 1. Frequency of seeing

Description

A flash of light of variable intensity was presented repeatedly at a fixed location in the visual field of a subject who reported whether the flash was visible.

Usage

data( "01_Miranda" )

Format

Data frame with three columns for 10 stimulus levels: x: Stimulus level. r: Number of successes. m: Number of trials.

References

Miranda, M. A. & Henson, D. B. ``Perimetric sensitivity and response variability in glaucoma with single-stimulus automated perimetry and multiple-stimulus perimetry with verbal feedback'', Acta Ophthalmologica, 86, 202-206, 2008.

[Package modelfree version 0.2 Index]