02_Levi {modelfree}	R Documentation

Example 2. Visual detection of path deviation

Description

The subject was presented with the image of a dot moving rightwards on a linear path until it reached the midline of the display, when it changed direction either upwards or downwards. The subject had to indicate the direction.

Usage

data( "02_Levi" )

Format

Data frame with three columns for 7 stimulus levels: x: Stimulus level. r: Number of successes. m: Number of trials.

References

Levi, D. M. & Tripathy, S. P. ``Is the ability to identify deviations in multiple trajectories compromised by amblyopia?'', Journal of Vision, 6(12), 1367-1379, 2006.

[Package modelfree version 0.2 Index]