03_Carcagno {modelfree}	R Documentation

Example 3. Discrimination of pitch

Description

The subject had to identify the interval containing a tone whose fundamental frequency was different from that in the other two intervals.

Usage

data( "03_Carcagno" )

Format

Data frame with three columns for 8 stimulus levels: x: Stimulus level. r: Number of successes. m: Number of trials.

References

Unpublished data from S. Carcagno, Lancaster University, July 2008

[Package modelfree version 0.2 Index]