04_Xie {modelfree}	R Documentation

Example 4. Discrimination of 'porthole' views of natural scenes

Description

The subject was presented with a display split into two parts, one containing a pair of patches from the same image, the other a pair from different images, and the subject had to judge which pair came from the same image.

Usage

data( "04_Xie" )

Format

Data frame with three columns for 10 stimulus levels: x: Stimulus level. r: Number of successes. m: Number of trials.

References

Xie, Y. & Griffin, L. D. ``A 'portholes' experiment for probing perception of small patches of natural images'', Perception, 36, 315, 2007.

[Package modelfree version 0.2 Index]