05_Schofield {modelfree}	R Documentation

Example 5. Induction of a visual motion aftereffect

Description

The subject was presented with a moving adaptation stimulus, followed by a test stimulus.

Usage

data( "05_Schofield" )

Format

Data frame with three columns for 7 stimulus levels: x: Stimulus level. r: Number of successes. m: Number of trials.

References

Schofield, A. J., Ledgeway, T., & Hutchinson, C. V. ``Asymmetric transfer of the dynamic motion aftereffect between first- and second-order cues and among different second-order cues'', Journal of Vision, 7(8), 1-12, 2007.

[Package modelfree version 0.2 Index]