06_Nascimento {modelfree}	R Documentation

Example 6. Discrimination of image approximations

Description

The subject was shown an image of a natural scene and an approximation of this image based on principal component analysis. The task was to distinguish between the images.

Usage

data( "06_Nascimento" )

Format

Data frame with three columns for 8 stimulus levels: x: Stimulus level. r: Number of successes. m: Number of trials.

References

Nascimento, S.M.C., Foster, D.H., & Amano, K. ``Psychophysical estimates of the number of spectral-reflectance basis functions needed to reproduce natural scenes'', Journal of the Optical Society of America A-Optics Image Science and Vision, 22 (6), 1017-1022, 2005.

[Package modelfree version 0.2 Index]