binom_gl {modelfree}	R Documentation

Psychometric function with guessing and lapsing rates

Description

THIS IS AN INTERNAL FUNCTION: USE BINOM_LIMS FOR BEST RESULTS. Maximum likelihood estimates of the parameters of psychometric function with guessing and lapsing rates (GLM) or only guessing rate. The estimated parameters for the linear part are in vector b and the estimated limits are in lims.

Usage


values<-binom_gl( r, m, x, link, p, K, initval );

Arguments

r 	 number of successes in points x

m 	 number of trials in points x

x 	 stimulus levels

link 	 link function

p 	 degree of the polynomial to be fitted on the linear scale

K 	 Power parameter in Weibull and reverse Weibull models

initval 	 initial value for guessing and lapsing rates

Value

value 	 Object with 4 components: guessing: estimated guessing rate lapsing: estimated lapsing rate b: vector of estimated coefficients for the linear part fit: glm object to be used in evaluation of fitted values

Examples


data( "01_Miranda" );
value <-binom_gl( example01$r, example01$m, example01$x, "logit", 1, 2, c( 0.01, 0.01 ) );

[Package modelfree version 0.2 Index]