binom_l {modelfree}	R Documentation

Psychometric function with lapsing rate

Description

THIS IS AN INTERNAL FUNCTION: USE BINOM_LIMS FOR BEST RESULTS. Maximum likelihood estimates of the parameters of psychometric function with lapsing rate (GLM). The estimated parameters for the linear part are in vector b and the estimated guessing rate is guess (GLM).

Usage


value<-binom_l( r, m, x, link, p, K, initval );

Arguments

r 	 number of successes in points x

m 	 number of trials in points x

x 	 stimulus levels

link 	 link function

p 	 degree of the polynomial to be fitted on the linear scale

K 	 Power parameter in Weibull and reverse Weibull models

initval 	 initial value for lapsing rate

Value

value 	 Object with 4 components: lapsing: estimated lapsing rate b: vector of estimated coefficients for the linear part fit: glm object to be used in evaluation of fitted values

Examples


data( "01_Miranda" );
value <-binom_l( example01$r, example01$m, example01$x, "logit", 1, 2, 0.01 );

[Package modelfree version 0.2 Index]