binom_revweib {modelfree}	R Documentation

Psychometric function fitting for reverse Weibull link function

Description

Maximum likelihood estimates of the parameters of the reverser Weibull model(GLM). The estimated parameters for the linear part are in vector b and the estimated exponent is K.

Usage


value<-binom_revweib( r, m, x, p = 1, initK = 2, guessing = 0, lapsing = 0 )

Arguments

r 	 number of successes in points x

m 	 number of trials in points x

x 	 stimulus levels

p 	 degree of the polynomial to be fitted on the linear scale; default is 1

initK 	 Power parameter in reverse Weibull model; default is 2

guessing 	 guessing rate; default is 0

lapsing 	 lapsing rate; default is 0

Value

value 	 Object with 3 components: b: vector of estimated coefficients for the linear part K: estiamte of the power parameter in the reverse Weibull model fit: glm object to be used in evaluation of fitted values

Examples


data( "01_Miranda" );
value <- binom_revweib( example01$r, example01$m, example01$x )

[Package modelfree version 0.2 Index]