binomfit_lims {modelfree}	R Documentation

Generalized linear model fit with guessing and lapsing rates

Description

The function fits a binomial generalised liner model with fixed guessing and lapsing rates

Usage


value <- binomfit_lims( r, m, x, p = 1, link = "logit", guessing = 0, lapsing = 0, K = 2 )

Arguments

r 	 number of successes in points x

m 	 number of trials in points x

x 	 stimulus levels

p 	 degree of the polynomial to be fitted on the linear scale; default is 1

link 	 link function; default is "logit"

guessing 	 guessing rate; default is 0

lapsing 	 lapsing rate; default is 0

Value

value 	 Object with 2 components: b: vector of estimated coefficients for the linear part fit: glm object to be used in evaluation of fitted values

Examples


data( "01_Miranda" );
value <- binomfit_lims( example01$r, example01$m, example01$x )

[Package modelfree version 0.2 Index]