bootstrap_ci_sl {modelfree}	R Documentation

Bootstrap estimate of confidence interval for slope estimation

Description

Finds bootstrap estimate of a confidence interval at a significant level alpha for the estimated slope for the local polynomial estimation of psychometric function (PF) with guessing and lapsing rates specified in lims. Confidence interval is based on bootstrap percentiles

Usage


value<-bootstrap_ci_sl( TH, r, m, x, N, h0, alpha = 0.05, X = (max(x)-min(x))*(0:999)/999+min(x), link = "logit", guessing = 0, lapsing = 0, K = 2, p = 1, ker = "dnorm", maxiter = 50, tol = 1e-6 );

Arguments

TH 	 required threshold level

r 	 number of successes in points x

m 	 number of trials in points x

x 	 stimulus levels

N 	 number of bootstrap replications

h0 	 pilot bandwidth; if not specified, then the scaled plug-in bandwidth is used

alpha 	 significance level of the confidence interval

X 	 set of value for which to calculate the estimates of PF for the thresholdestimation; if not given 1000 equally spaced points from min to max of xdes are used

link 	 name of the link function to be used; default is "logit"

guessing 	 guessing rate; default is 0

lapsing 	 lapsing rate; default is 0

K 	 power parameter for Weibull and reverse Weibull link; default is 2

p 	 order of the polynomial; default is 1

ker 	 kernel function for weights; default "dnorm"

maxiter 	 maximum number of iterations in Fisher scoring; default is 50

tol 	 tolerance level at which to stop Fisher scoring; default is 1e-6

Value

value 	 Object with 2 components: ci: confidence interval based on bootstrap percentiles sl0: slope estimate

Examples


data( "01_Miranda" )
bwd <- 0.2959;
value <- bootstrap_ci_sl( 0.5, example01$r, example01$m, example01$x, 10, bwd );

[Package modelfree version 0.2 Index]