checkinput {modelfree}	R Documentation

Check input parameters of modelfree functions

Description

THIS IS AN INTERNAL FUNCTION ONLY. Pool of routines that check robustness of input of parameters passed to the other functions of modelfree package.

Usage

checkinput(type,x)

Arguments

type 	 Type of checking

x 	 Input data to be checked

Value

None. Only stops execution if data is not consistent.

[Package modelfree version 0.2 Index]