comploglog_link_private {modelfree}	R Documentation

Complementary log-log link function with guessing and lapsing rates

Description

THIS IS AN INTERNAL FUNCTION: USE COMPLOGLOG_LINK FOR BEST RESULTS. Complementaty loglog link for use with GLM functions. The guessing and lapsing rate are fixed to values given in lims, hence link is a function of only one variable

Usage


link<-comploglog_link_private( guessing = 0, lapsing = 0 );

Arguments

guessing 	 guessing rate; default is 0

lapsing 	 lapsing rate; default is 0

Value

link 	 Complementary log-log link for use in all GLM functions

Examples


data( "01_Miranda" );
x <- example01$x;
r <- example01$r;
m <- example01$m;
glmdata <- data.frame( cbind( r/m ,m , x ) );
names( glmdata ) <- c( "resp", "m", "x" );
glmformula <- c( "resp ~ x" );
userlink<-comploglog_link_private( 0.1, 0.1 );
fit <- glm( glmformula, data = glmdata, weights = m, family = binomial( userlink ) );

[Package modelfree version 0.2 Index]