revweibull_link {modelfree}	R Documentation

Reverse Weibull link function with guessing and lapsing rates

Description

Reverse Weibull link for use with GLM functions. The guessing rate and lapsing rate are fixed, and power parameter is set to be equal K, hence link is a function of only one variable

Usage


link<-revweibull_link( K, guessing = 0, lapsing = 0 );

Arguments

K 	 power parameter for reverse Weibull link function

guessing 	 guessing rate; default is 0

lapsing 	 lapsing rate; default is 0

Value

link 	 Reverse Weibull link for use in all GLM functions

Examples


data( "01_Miranda" );
x <- example01$x;
r <- example01$r;
m <- example01$m;
glmdata <- data.frame( cbind( r/m ,m , x ) );
names( glmdata ) <- c( "resp", "m", "x" );
glmformula <- c( "resp ~ x" );
userlink<-revweibull_link( 20 );
fit <- glm( glmformula, data = glmdata, weights = m, family = binomial( userlink ) );

[Package modelfree version 0.2 Index]