threshold_slope {modelfree}	R Documentation

Threshold and slope of estimated psychometric function

Description

Finds the approximate value of x (=x_th) for which the value of the psychometric function is equal thresh and the approximate value of slope in x_th

Usage


value<-threshold_slope( pfit, xfit, thresh = 0.5 );

Arguments

pfit 	 estimated values of the psychometric function

xfit 	 stimulus levels in which the function was estimated

thresh 	 value for which to estimate threshold; default is 0.5

Value

value 	 Object with 2 elements: x_th: estimated threshold slope: estimated value of slope, i.e. derivative of pfit at x_th

Examples


data( "01_Miranda" );
xnew = 1.2 * (0:999)/999+0.1;
h = 0.2959;
fit <- locglmfit( xnew, example01$r, example01$m, example01$x, h );
value <- threshold_slope( fit$pfit , xnew );

[Package modelfree version 0.2 Index]